backend/requirements.txt
numpy
flask
flask-cors
transformers>=4.34.0
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision 
torchaudio
--index-url https://pypi.org/simple
bitsandbytes>=0.41.0
accelerate
sentencepiece
protobuf
# Update llama-cpp-python for better CUDA 12.8 and TensorRT 10.x compatibility
llama-cpp-python>=0.2.38
pycuda>=2023.1
tensorrt>=10.0.0
